split up several modules into sub packages
