+++ src/tables/cli/args.py
@@ -17,8 +17,6 @@ Each command has its own seperate set of flags/arguments.
 from __future__ import annotations
 import argparse
 from tables.domain.enums import CliCommands
-from tables.domain.models import DatabaseConnection
-from tables import serializers
 
 class CliArgs:
 
@@ -89,19 +87,5 @@ class CliArgs:
             command = CliCommands.LIST
 
         return command
-
-
-def get_database_connection(cli_args: CliArgs) -> DatabaseConnection:
-    """Creat a new DatabaseConnection model with property values provided in the CLI args."""
-
-    args_dict = cli_args.args.__dict__
-    return serializers.to_database_connection(args_dict)
-
-
-
-
-
-    
-
