+++ src/tables/domain/constants.py
