+++ src/tables/domain/errors.py
