+++ src/tables/persistence/json_encoder.py
